feat: started implementing drag and drop for user magnets
