+++ src/components/board/UserList.tsx
@@ -1,3 +1,5 @@
+import { Draggable, Droppable, DroppableProvided } from "@hello-pangea/dnd";
+
 import { User } from "@/types";
 
 import UserMagnet from "./UserMagnet";
@@ -8,16 +10,39 @@ interface UserListProps {
 
 const UserList: React.FC<UserListProps> = ({ users }) => {
     return (
-        <div style={{
-            display: "flex",
-            alignItems: "center",
-            border: "solid 1px black",
-            borderRadius: "10px",
-        }}>
-            {users.map((user) => (
-                <UserMagnet key={user.userid} user={user} />
-            ))}
-        </div>
+        <Droppable droppableId="user-list" type="user" direction="horizontal">
+            {(provided: DroppableProvided) => {
+                return(
+                    <div
+                        ref={provided.innerRef}
+                        {...provided.droppableProps}
+                        style={{ display: "flex",
+                        alignItems: "center",
+                        border: "solid 1px black",
+                        borderRadius: "10px",
+                        }}
+                    >
+                        {users && users.map((user, index) => (
+                            <Draggable key={user.userid} draggableId={user.userid} index={index}>
+                                {(provided) => {
+                                    return (
+                                        <div
+                                            ref={provided.innerRef}
+                                            {...provided.draggableProps}
+                                            {...provided.dragHandleProps}
+                                        >
+                                            <UserMagnet user={user} />
+                                        </div>
+                                    );
+                                }}
+                            </Draggable>
+                        ))}
+                        {provided.placeholder}
+                    </div>
+                );
+
+            }}
+        </Droppable>
     );
 };
