+++ src/components/board/UserMagnet.tsx
@@ -8,12 +8,11 @@ interface UserMagnetProps {
 
 const UserMagnet: React.FC<UserMagnetProps> = ({ user }) => {
 
-    //make magnet styles
-    //make droppable
     return (
         <div style={{
             display: "flex",
             justifyContent: "center",
+            backgroundColor: user.color || "pink",
             alignItems: "center", margin: "5px", border: "solid 1px", borderRadius: "10px", color: "black", width: "60px", overflow: "hidden", height:"20px"
         }}>
             <Typography sx={{
